feat: move all modal files into a separate folder.
