+++ client/src/components/RatingsReviews/ReviewTile.jsx
@@ -8,7 +8,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCheck } from '@fortawesome/free-solid-svg-icons';
 
 import RatingsReviewsCSS from './RatingsReviews.module.css';
-import ReviewImageModal from './ReviewImageModal.jsx';
+import ReviewImageModal from './Modals/ReviewImageModal.jsx';
 
 export default class ReviewTile extends React.Component {
   constructor(props) {
